--- go_lib/dependency/extenders/main.go
@@ -1,32 +1,0 @@
-/*
-Copyright 2024 Flant JSC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package extenders
-
-import (
-	"slices"
-
-	"github.com/deckhouse/deckhouse/deckhouse-controller/pkg/apis/deckhouse.io/v1alpha1"
-)
-
-func IsExtendersField(field string) bool {
-	return slices.Contains([]string{
-		v1alpha1.KubernetesRequirementFieldName,
-		v1alpha1.DeckhouseRequirementFieldName,
-		v1alpha1.BootstrappedRequirementFieldName,
-		v1alpha1.ModuleDependencyRequirementFieldName,
-	}, field)
-}
